use a distinct random generator for scene textures
Using the same random generator as the one use by the STM
would cause some conditions to be executed less often than
in the original game.

This was for example noticeable when starting the game, the
"order to activate security drones" message would not be played
most of the time. The corresponding conditions relied on a
rnd(5) call.

SCRIPT 399 OFFSET 0x8280
        CONDITIONS SEQUENCE
                get_dist2 ( 0, OBJ.CUSTOM[0] )
                test_val ( VAR_dist2, 4, OBJ.CUSTOM[4] )
                rnd ( 5 )
        STATEMENTS SEQUENCE
                msg ( 1, 152 )