Add current datafiles voice setting to the savegame data
Spanish and Italian have different .dtt data depending on the voice
setting. German, French and English are not affected.
Ideally, the savegame data should be made language independent but
GameObject.text still points to the .dtt data, the information to
look that information would have to be persisted (or the string data
be actually strdup'ed)